+++ tests/trezor/trezor_tests.cpp
@@ -1479,7 +1479,6 @@ tsx_builder * tsx_builder::construct_pending_tx(tools::wallet2::pending_tx &ptx,
   ptx.tx_key = tx_key;
   ptx.additional_tx_keys = additional_tx_keys;
   ptx.dests = m_destinations;
-  ptx.multisig_sigs.clear();
   ptx.construction_data.sources = m_sources;
   ptx.construction_data.change_dts = m_destinations.back();
   ptx.construction_data.splitted_dsts = m_destinations;